fix(template): keep managed categories server-only
