+++ frontend/providers/template/src/pages/api/v2alpha/templates/index.ts
@@ -1,5 +1,6 @@
 import { TemplateType } from '@/types/app';
-import { getCategorySlugs, getConfiguredTemplateCategories } from '@/utils/template';
+import { getCategorySlugs } from '@/utils/template';
+import { getConfiguredTemplateCategories } from '@/utils/templateCategories.server';
 import type { NextApiRequest, NextApiResponse } from 'next';
 import path from 'path';
 import { getCachedTemplates } from './templateCache';
